fix(ContextMenuCanaryy): fix all mistakes
closes #1929
